\name{geco.imageCol}
\alias{geco.imageCol}
\title{geco.imageCol}
\description{Function to make a visual representation of a matrix containing color codes}
\usage{geco.imageCol(matcol=NULL,						  strat=NULL,						  xlab.cex=0.5,						  ylab.cex=0,						  ...						  )}
\arguments{
\item{matcol}{Color matrix}
\item{strat}{Column of matrix used to draw split lines}
\item{xlab}{Size of x axis labels (if 0 no labels are drawn)}
\item{ylab}{Size of y axis labels (if 0 no labels are drawn)}
}
\value{None
}
\author{Eric Letouze}
\keyword{methods}
