\name{geco.imagePval}
\alias{geco.imagePval}
\title{geco.imagePval}
\description{Function to produce a graphical representation of a pvalue table}
\usage{geco.imagePval(pval=NULL,...)}
\arguments{
\item{pval}{P-value table}
\item{...}{required}
}
\value{None
}
\author{Eric Letouze}
\keyword{methods}
